refactor root.preRun to support composition
The root.preRun func assigned to ```rootCmd.PersistentPreRunE``` relies
on Project.Open(), which assumes all child commands will want to operate
on an existing Project.

While it's no big deal to assign root.PreRunNoop, that actually seems to
mess with a couple of the truly shared global flag handlers.
For example, ```log-level```, ```chdir``` and ```dry-run``` are handled in the
```root.preRun``` right now. This makes the ```PreRunNoop``` trick come
with the odd side-effect of those flags being unhandled.

The Cobra ```command.SetGlobalNormalizationFunc``` might be abused to handle
flag-triggered actions, but that seems dangerous as well. :skull:

As a starting point, this PR separates out the ```Project.Open()``` and the other
calls, which makes it easier to compose a custom ```PersistentPreRunE``` in
child commands that don't assume a Project already exists.

This is necessary for the work I'm doing today on #229 (which will come in a different PR).

Any better ways of handling this would be awesome. Decomposing the flag-triggered stuff
even further might be good to consider.