Adds a flag for passing an AWS credentials file into Apex
This is rebased on a cleaner apex/apex master.

Fixes AWS credentials short flag conflict

There was a crasher with the ```-c``` flag of ```apex deploy``` because a prior version of this
hijacked that short flag. For now, it's just removed and the function is changed to
```StringVar``` from ```StringVarP```